refactor: centralize attachment HTTP logic in attachment.Service
Introduce attachment.Service as a spath-agnostic HTTP layer exposing
List, Remove, Download, Upload, and Attach. Remove the domain-specific
IssueService, WikiService, PullRequestService, and SpaceService types
from internal/attachment.

Add domain wrapper AttachmentService to internal/wiki, internal/issue,
internal/pullrequest, and internal/space. Each wrapper is responsible
only for validation and spath construction; all HTTP calls are
delegated to the shared attachment.Service.

Update root-package constructors in wiki.go, issue.go, pullrequest.go,
and space.go to use the new domain wrappers.

Move attachment service tests from internal/attachment to the
respective domain packages (internal/wiki, internal/issue,
internal/pullrequest). Update internal/attachment tests to cover only
the generic Service methods (Upload, List, Remove, Download, Attach).

Part of #281
