+++ internal/attachment/service.go
@@ -5,16 +5,22 @@ import (
 	"errors"
 	"io"
 	"net/url"
-	"path"
 	"strconv"
 
 	"github.com/nattokin/go-backlog/internal/core"
 	"github.com/nattokin/go-backlog/internal/model"
-	"github.com/nattokin/go-backlog/internal/validate"
 )
 
-func ListAttachments(ctx context.Context, m *core.Method, spath string) ([]*model.Attachment, error) {
-	resp, err := m.Get(ctx, spath, nil)
+// Service holds shared HTTP logic for attachment-related Backlog API endpoints.
+// It is spath-agnostic: callers supply the full sub-path and are responsible
+// for validation and path construction.
+type Service struct {
+	method *core.Method
+}
+
+// List returns the list of attachments at spath.
+func (s *Service) List(ctx context.Context, spath string) ([]*model.Attachment, error) {
+	resp, err := s.method.Get(ctx, spath, nil)
 	if err != nil {
 		return nil, err
 	}
@@ -27,8 +33,9 @@ func ListAttachments(ctx context.Context, m *core.Method, spath string) ([]*mode
 	return v, nil
 }
 
-func RemoveAttachment(ctx context.Context, m *core.Method, spath string) (*model.Attachment, error) {
-	resp, err := m.Delete(ctx, spath, nil)
+// Remove deletes the attachment at spath and returns the deleted attachment.
+func (s *Service) Remove(ctx context.Context, spath string) (*model.Attachment, error) {
+	resp, err := s.method.Delete(ctx, spath, nil)
 	if err != nil {
 		return nil, err
 	}
@@ -41,44 +48,8 @@ func RemoveAttachment(ctx context.Context, m *core.Method, spath string) (*model
 	return v, nil
 }
 
-// ──────────────────────────────────────────────────────────────
-//  IssueService
-// ──────────────────────────────────────────────────────────────
-
-type IssueService struct {
-	method *core.Method
-}
-
-func (s *IssueService) List(ctx context.Context, issueIDOrKey string) ([]*model.Attachment, error) {
-	if err := validate.ValidateIssueIDOrKey(issueIDOrKey); err != nil {
-		return nil, err
-	}
-
-	spath := path.Join("issues", issueIDOrKey, "attachments")
-	return ListAttachments(ctx, s.method, spath)
-}
-
-func (s *IssueService) Remove(ctx context.Context, issueIDOrKey string, attachmentID int) (*model.Attachment, error) {
-	if err := validate.ValidateIssueIDOrKey(issueIDOrKey); err != nil {
-		return nil, err
-	}
-	if err := validate.ValidateAttachmentID(attachmentID); err != nil {
-		return nil, err
-	}
-
-	spath := path.Join("issues", issueIDOrKey, "attachments", strconv.Itoa(attachmentID))
-	return RemoveAttachment(ctx, s.method, spath)
-}
-
-func (s *IssueService) Download(ctx context.Context, issueIDOrKey string, attachmentID int) (*model.FileData, error) {
-	if err := validate.ValidateIssueIDOrKey(issueIDOrKey); err != nil {
-		return nil, err
-	}
-	if err := validate.ValidateAttachmentID(attachmentID); err != nil {
-		return nil, err
-	}
-
-	spath := path.Join("issues", issueIDOrKey, "attachments", strconv.Itoa(attachmentID))
+// Download streams the attachment at spath.
+func (s *Service) Download(ctx context.Context, spath string) (*model.FileData, error) {
 	resp, err := s.method.Download(ctx, spath, nil)
 	if err != nil {
 		return nil, err
@@ -87,80 +58,9 @@ func (s *IssueService) Download(ctx context.Context, issueIDOrKey string, attach
 	return core.DownloadResponse(resp)
 }
 
-// ──────────────────────────────────────────────────────────────
-//  PullRequestService
-// ──────────────────────────────────────────────────────────────
-
-type PullRequestService struct {
-	method *core.Method
-}
-
-func (s *PullRequestService) List(ctx context.Context, projectIDOrKey string, repositoryIDOrName string, prNumber int) ([]*model.Attachment, error) {
-	if err := validate.ValidateProjectIDOrKey(projectIDOrKey); err != nil {
-		return nil, err
-	}
-	if err := validate.ValidateRepositoryIDOrName(repositoryIDOrName); err != nil {
-		return nil, err
-	}
-	if err := validate.ValidatePRNumber(prNumber); err != nil {
-		return nil, err
-	}
-
-	spath := path.Join("projects", projectIDOrKey, "git", "repositories", repositoryIDOrName, "pullRequests", strconv.Itoa(prNumber), "attachments")
-	return ListAttachments(ctx, s.method, spath)
-}
-
-func (s *PullRequestService) Remove(ctx context.Context, projectIDOrKey string, repositoryIDOrName string, prNumber int, attachmentID int) (*model.Attachment, error) {
-	if err := validate.ValidateProjectIDOrKey(projectIDOrKey); err != nil {
-		return nil, err
-	}
-	if err := validate.ValidateRepositoryIDOrName(repositoryIDOrName); err != nil {
-		return nil, err
-	}
-	if err := validate.ValidatePRNumber(prNumber); err != nil {
-		return nil, err
-	}
-	if err := validate.ValidateAttachmentID(attachmentID); err != nil {
-		return nil, err
-	}
-
-	spath := path.Join("projects", projectIDOrKey, "git", "repositories", repositoryIDOrName, "pullRequests", strconv.Itoa(prNumber), "attachments", strconv.Itoa(attachmentID))
-	return RemoveAttachment(ctx, s.method, spath)
-}
-
-func (s *PullRequestService) Download(ctx context.Context, projectIDOrKey string, repositoryIDOrName string, prNumber int, attachmentID int) (*model.FileData, error) {
-	if err := validate.ValidateProjectIDOrKey(projectIDOrKey); err != nil {
-		return nil, err
-	}
-	if err := validate.ValidateRepositoryIDOrName(repositoryIDOrName); err != nil {
-		return nil, err
-	}
-	if err := validate.ValidatePRNumber(prNumber); err != nil {
-		return nil, err
-	}
-	if err := validate.ValidateAttachmentID(attachmentID); err != nil {
-		return nil, err
-	}
-
-	spath := path.Join("projects", projectIDOrKey, "git", "repositories", repositoryIDOrName, "pullRequests", strconv.Itoa(prNumber), "attachments", strconv.Itoa(attachmentID))
-	resp, err := s.method.Download(ctx, spath, nil)
-	if err != nil {
-		return nil, err
-	}
-
-	return core.DownloadResponse(resp)
-}
-
-// ──────────────────────────────────────────────────────────────
-//  SpaceService
-// ──────────────────────────────────────────────────────────────
-
-type SpaceService struct {
-	method *core.Method
-}
-
-func (s *SpaceService) Upload(ctx context.Context, fileName string, r io.Reader) (*model.Attachment, error) {
-	resp, err := s.method.Upload(ctx, "space/attachment", fileName, r)
+// Upload uploads a file to spath.
+func (s *Service) Upload(ctx context.Context, spath string, fileName string, r io.Reader) (*model.Attachment, error) {
+	resp, err := s.method.Upload(ctx, spath, fileName, r)
 	if err != nil {
 		return nil, err
 	}
@@ -173,31 +73,21 @@ func (s *SpaceService) Upload(ctx context.Context, fileName string, r io.Reader)
 	return &v, nil
 }
 
-// ──────────────────────────────────────────────────────────────
-//  WikiService
-// ──────────────────────────────────────────────────────────────
-
-type WikiService struct {
-	method *core.Method
-}
-
-func (s *WikiService) Attach(ctx context.Context, wikiID int, attachmentIDs []int) ([]*model.Attachment, error) {
-	if err := validate.ValidateWikiID(wikiID); err != nil {
-		return nil, err
-	}
+// Attach links already-uploaded space attachments to a resource at spath.
+// attachmentIDs must not be empty and every ID must be positive.
+func (s *Service) Attach(ctx context.Context, spath string, attachmentIDs []int) ([]*model.Attachment, error) {
 	if len(attachmentIDs) == 0 {
 		return nil, errors.New("attachmentIDs must not be empty")
 	}
 
 	form := url.Values{}
 	for _, id := range attachmentIDs {
-		if err := validate.ValidateAttachmentID(id); err != nil {
-			return nil, err
+		if id <= 0 {
+			return nil, errors.New("attachmentID must be greater than 0")
 		}
 		form.Add("attachmentId[]", strconv.Itoa(id))
 	}
 
-	spath := path.Join("wikis/", strconv.Itoa(wikiID), "/attachments")
 	resp, err := s.method.Post(ctx, spath, form)
 	if err != nil {
 		return nil, err
@@ -211,68 +101,11 @@ func (s *WikiService) Attach(ctx context.Context, wikiID int, attachmentIDs []in
 	return v, nil
 }
 
-func (s *WikiService) List(ctx context.Context, wikiID int) ([]*model.Attachment, error) {
-	if err := validate.ValidateWikiID(wikiID); err != nil {
-		return nil, err
-	}
-
-	spath := path.Join("wikis", strconv.Itoa(wikiID), "attachments")
-	return ListAttachments(ctx, s.method, spath)
-}
-
-func (s *WikiService) Remove(ctx context.Context, wikiID, attachmentID int) (*model.Attachment, error) {
-	if err := validate.ValidateWikiID(wikiID); err != nil {
-		return nil, err
-	}
-	if err := validate.ValidateAttachmentID(attachmentID); err != nil {
-		return nil, err
-	}
-
-	spath := path.Join("wikis", strconv.Itoa(wikiID), "attachments", strconv.Itoa(attachmentID))
-	return RemoveAttachment(ctx, s.method, spath)
-}
-
-func (s *WikiService) Download(ctx context.Context, wikiID, attachmentID int) (*model.FileData, error) {
-	if err := validate.ValidateWikiID(wikiID); err != nil {
-		return nil, err
-	}
-	if err := validate.ValidateAttachmentID(attachmentID); err != nil {
-		return nil, err
-	}
-
-	spath := path.Join("wikis", strconv.Itoa(wikiID), "attachments", strconv.Itoa(attachmentID))
-	resp, err := s.method.Download(ctx, spath, nil)
-	if err != nil {
-		return nil, err
-	}
-
-	return core.DownloadResponse(resp)
-}
-
 // ──────────────────────────────────────────────────────────────
-//  Constructors
+//  Constructor
 // ──────────────────────────────────────────────────────────────
 
-func NewPullRequestService(method *core.Method) *PullRequestService {
-	return &PullRequestService{
-		method: method,
-	}
-}
-
-func NewIssueService(method *core.Method) *IssueService {
-	return &IssueService{
-		method: method,
-	}
-}
-
-func NewSpaceService(method *core.Method) *SpaceService {
-	return &SpaceService{
-		method: method,
-	}
-}
-
-func NewWikiService(method *core.Method) *WikiService {
-	return &WikiService{
-		method: method,
-	}
+// NewService creates and returns a new attachment Service.
+func NewService(method *core.Method) *Service {
+	return &Service{method: method}
 }
