--- internal/attachment/service_wiki_test.go
@@ -1,291 +1,0 @@
-package attachment_test
-
-import (
-	"context"
-	"errors"
-	"net/http"
-	"net/url"
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-
-	"github.com/nattokin/go-backlog/internal/attachment"
-	"github.com/nattokin/go-backlog/internal/model"
-	"github.com/nattokin/go-backlog/internal/testutil/fixture"
-	"github.com/nattokin/go-backlog/internal/testutil/mock"
-)
-
-func TestWikiAttachmentService_Attach(t *testing.T) {
-	cases := map[string]struct {
-		wikiID        int
-		attachmentIDs []int
-
-		expectError bool
-		want        []*model.Attachment
-
-		mockPostFn func(ctx context.Context, spath string, form url.Values) (*http.Response, error)
-	}{
-		"success-single": {
-			wikiID:        1234,
-			attachmentIDs: []int{2},
-			want:          newTestAttachmentSingleList(),
-			mockPostFn: func(ctx context.Context, spath string, form url.Values) (*http.Response, error) {
-				assert.Equal(t, "wikis/1234/attachments", spath)
-				v := form
-				assert.Equal(t, []string{"2"}, v["attachmentId[]"])
-				return mock.NewJSONResponse(fixture.Attachment.SingleListJSON), nil
-			},
-		},
-
-		"success-multiple": {
-			wikiID:        1,
-			attachmentIDs: []int{2, 5},
-			want:          newTestAttachmentList(),
-			mockPostFn: func(ctx context.Context, spath string, form url.Values) (*http.Response, error) {
-				return mock.NewJSONResponse(fixture.Attachment.ListJSON), nil
-			},
-		},
-
-		"error-wikiID-invalid": {
-			wikiID:        0,
-			attachmentIDs: []int{1, 2},
-			expectError:   true,
-			mockPostFn:    mock.NewUnexpectedPostFn(t),
-		},
-
-		"error-attachmentIDs-invalid": {
-			wikiID:        1,
-			attachmentIDs: []int{0, 1, 2},
-			expectError:   true,
-			mockPostFn:    mock.NewUnexpectedPostFn(t),
-		},
-
-		"error-attachmentIDs-empty": {
-			wikiID:        1,
-			attachmentIDs: []int{},
-			expectError:   true,
-			mockPostFn:    mock.NewUnexpectedPostFn(t),
-		},
-
-		"error-client": {
-			wikiID:        1234,
-			attachmentIDs: []int{2},
-			expectError:   true,
-			mockPostFn: func(ctx context.Context, spath string, form url.Values) (*http.Response, error) {
-				return nil, errors.New("error")
-			},
-		},
-
-		"error-invalid-json": {
-			wikiID:        1234,
-			attachmentIDs: []int{2},
-			expectError:   true,
-			mockPostFn: func(ctx context.Context, spath string, form url.Values) (*http.Response, error) {
-				return mock.NewJSONResponse(fixture.InvalidJSON), nil
-			},
-		},
-	}
-
-	for name, tc := range cases {
-		t.Run(name, func(t *testing.T) {
-			t.Parallel()
-
-			method := mock.NewMethod(t)
-			method.Post = tc.mockPostFn
-			s := attachment.NewWikiService(method)
-
-			attachments, err := s.Attach(context.Background(), tc.wikiID, tc.attachmentIDs)
-
-			if tc.expectError {
-				assert.Error(t, err)
-				assert.Nil(t, attachments)
-				return
-			}
-
-			assert.NoError(t, err)
-			require.NotNil(t, attachments)
-
-			assert.Len(t, attachments, len(tc.want))
-
-			for i, w := range tc.want {
-				assert.Equal(t, w.ID, attachments[i].ID)
-				assert.Equal(t, w.Name, attachments[i].Name)
-				assert.Equal(t, w.Size, attachments[i].Size)
-				assert.ObjectsAreEqualValues(w.Created, attachments[i].Created)
-			}
-		})
-	}
-}
-
-func TestWikiAttachmentService_List(t *testing.T) {
-	cases := map[string]struct {
-		wikiID int
-
-		expectError bool
-		want        []*model.Attachment
-
-		mockGetFn func(ctx context.Context, spath string, query url.Values) (*http.Response, error)
-	}{
-		"success": {
-			wikiID: 1234,
-			want:   newTestAttachmentList(),
-			mockGetFn: func(ctx context.Context, spath string, query url.Values) (*http.Response, error) {
-				assert.Equal(t, "wikis/1234/attachments", spath)
-				return mock.NewJSONResponse(fixture.Attachment.ListJSON), nil
-			},
-		},
-
-		"error-wikiID-zero": {
-			wikiID:      0,
-			expectError: true,
-			mockGetFn:   mock.NewUnexpectedGetFn(t),
-		},
-
-		"error-wikiID-negative": {
-			wikiID:      -1,
-			expectError: true,
-			mockGetFn:   mock.NewUnexpectedGetFn(t),
-		},
-
-		"error-client": {
-			wikiID:      1234,
-			expectError: true,
-			mockGetFn: func(ctx context.Context, spath string, query url.Values) (*http.Response, error) {
-				return nil, errors.New("error")
-			},
-		},
-
-		"error-invalid-json": {
-			wikiID:      1234,
-			expectError: true,
-			mockGetFn: func(ctx context.Context, spath string, query url.Values) (*http.Response, error) {
-				return mock.NewJSONResponse(fixture.InvalidJSON), nil
-			},
-		},
-	}
-
-	for name, tc := range cases {
-		t.Run(name, func(t *testing.T) {
-			t.Parallel()
-
-			method := mock.NewMethod(t)
-			method.Get = tc.mockGetFn
-			s := attachment.NewWikiService(method)
-
-			attachments, err := s.List(context.Background(), tc.wikiID)
-
-			if tc.expectError {
-				assert.Error(t, err)
-				assert.Nil(t, attachments)
-				return
-			}
-
-			assert.NoError(t, err)
-			require.NotNil(t, attachments)
-
-			assert.Len(t, attachments, len(tc.want))
-
-			for i, w := range tc.want {
-				assert.Equal(t, w.ID, attachments[i].ID)
-				assert.Equal(t, w.Name, attachments[i].Name)
-				assert.Equal(t, w.Size, attachments[i].Size)
-				assert.ObjectsAreEqualValues(w.Created, attachments[i].Created)
-			}
-		})
-	}
-}
-
-func TestWikiAttachmentService_Remove(t *testing.T) {
-	cases := map[string]struct {
-		wikiID       int
-		attachmentID int
-
-		expectError bool
-		want        *model.Attachment
-
-		mockDeleteFn func(ctx context.Context, spath string, form url.Values) (*http.Response, error)
-	}{
-		"success": {
-			wikiID:       1234,
-			attachmentID: 8,
-			want:         newTestAttachment(),
-			mockDeleteFn: func(ctx context.Context, spath string, form url.Values) (*http.Response, error) {
-				assert.Equal(t, "wikis/1234/attachments/8", spath)
-				return mock.NewJSONResponse(fixture.Attachment.SingleJSON), nil
-			},
-		},
-
-		"error-wikiID-zero": {
-			wikiID:       0,
-			attachmentID: 8,
-			expectError:  true,
-			mockDeleteFn: mock.NewUnexpectedDeleteFn(t),
-		},
-
-		"error-wikiID-negative": {
-			wikiID:       -1,
-			attachmentID: 8,
-			expectError:  true,
-			mockDeleteFn: mock.NewUnexpectedDeleteFn(t),
-		},
-
-		"error-attachmentID-zero": {
-			wikiID:       1,
-			attachmentID: 0,
-			expectError:  true,
-			mockDeleteFn: mock.NewUnexpectedDeleteFn(t),
-		},
-
-		"error-attachmentID-negative": {
-			wikiID:       1,
-			attachmentID: -1,
-			expectError:  true,
-			mockDeleteFn: mock.NewUnexpectedDeleteFn(t),
-		},
-
-		"error-client": {
-			wikiID:       1234,
-			attachmentID: 8,
-			expectError:  true,
-			mockDeleteFn: func(ctx context.Context, spath string, form url.Values) (*http.Response, error) {
-				return nil, errors.New("error")
-			},
-		},
-
-		"error-invalid-json": {
-			wikiID:       1234,
-			attachmentID: 8,
-			expectError:  true,
-			mockDeleteFn: func(ctx context.Context, spath string, form url.Values) (*http.Response, error) {
-				return mock.NewJSONResponse(fixture.InvalidJSON), nil
-			},
-		},
-	}
-
-	for name, tc := range cases {
-		t.Run(name, func(t *testing.T) {
-			t.Parallel()
-
-			method := mock.NewMethod(t)
-			method.Delete = tc.mockDeleteFn
-			s := attachment.NewWikiService(method)
-
-			attachment, err := s.Remove(context.Background(), tc.wikiID, tc.attachmentID)
-
-			if tc.expectError {
-				assert.Error(t, err)
-				assert.Nil(t, attachment)
-				return
-			}
-
-			assert.NoError(t, err)
-			require.NotNil(t, attachment)
-
-			assert.Equal(t, tc.want.ID, attachment.ID)
-			assert.Equal(t, tc.want.Name, attachment.Name)
-			assert.Equal(t, tc.want.Size, attachment.Size)
-			assert.ObjectsAreEqualValues(tc.want.Created, attachment.Created)
-		})
-	}
-}
