+++ internal/pullrequest/service_attachment.go
@@ -1,0 +1,89 @@
+package pullrequest
+
+import (
+	"context"
+	"path"
+	"strconv"
+
+	"github.com/nattokin/go-backlog/internal/attachment"
+	"github.com/nattokin/go-backlog/internal/core"
+	"github.com/nattokin/go-backlog/internal/model"
+	"github.com/nattokin/go-backlog/internal/validate"
+)
+
+// AttachmentService handles attachment-related Backlog API calls for pull requests.
+// It delegates all HTTP operations to the shared attachment.Service and is
+// responsible only for validation and spath construction.
+type AttachmentService struct {
+	svc *attachment.Service
+}
+
+// List returns a list of all attachments in the pull request.
+//
+// Backlog API docs: https://developer.nulab.com/docs/backlog/api/2/get-list-of-pull-request-attachment
+func (s *AttachmentService) List(ctx context.Context, projectIDOrKey string, repositoryIDOrName string, prNumber int) ([]*model.Attachment, error) {
+	if err := validate.ValidateProjectIDOrKey(projectIDOrKey); err != nil {
+		return nil, err
+	}
+	if err := validate.ValidateRepositoryIDOrName(repositoryIDOrName); err != nil {
+		return nil, err
+	}
+	if err := validate.ValidatePRNumber(prNumber); err != nil {
+		return nil, err
+	}
+
+	spath := path.Join("projects", projectIDOrKey, "git", "repositories", repositoryIDOrName, "pullRequests", strconv.Itoa(prNumber), "attachments")
+	return s.svc.List(ctx, spath)
+}
+
+// Remove removes a file attached to the pull request.
+//
+// Backlog API docs: https://developer.nulab.com/docs/backlog/api/2/delete-pull-request-attachments
+func (s *AttachmentService) Remove(ctx context.Context, projectIDOrKey string, repositoryIDOrName string, prNumber int, attachmentID int) (*model.Attachment, error) {
+	if err := validate.ValidateProjectIDOrKey(projectIDOrKey); err != nil {
+		return nil, err
+	}
+	if err := validate.ValidateRepositoryIDOrName(repositoryIDOrName); err != nil {
+		return nil, err
+	}
+	if err := validate.ValidatePRNumber(prNumber); err != nil {
+		return nil, err
+	}
+	if err := validate.ValidateAttachmentID(attachmentID); err != nil {
+		return nil, err
+	}
+
+	spath := path.Join("projects", projectIDOrKey, "git", "repositories", repositoryIDOrName, "pullRequests", strconv.Itoa(prNumber), "attachments", strconv.Itoa(attachmentID))
+	return s.svc.Remove(ctx, spath)
+}
+
+// Download downloads a file attached to the pull request.
+// The caller is responsible for closing FileData.Body after use.
+//
+// Backlog API docs: https://developer.nulab.com/docs/backlog/api/2/download-pull-request-attachment
+func (s *AttachmentService) Download(ctx context.Context, projectIDOrKey string, repositoryIDOrName string, prNumber int, attachmentID int) (*model.FileData, error) {
+	if err := validate.ValidateProjectIDOrKey(projectIDOrKey); err != nil {
+		return nil, err
+	}
+	if err := validate.ValidateRepositoryIDOrName(repositoryIDOrName); err != nil {
+		return nil, err
+	}
+	if err := validate.ValidatePRNumber(prNumber); err != nil {
+		return nil, err
+	}
+	if err := validate.ValidateAttachmentID(attachmentID); err != nil {
+		return nil, err
+	}
+
+	spath := path.Join("projects", projectIDOrKey, "git", "repositories", repositoryIDOrName, "pullRequests", strconv.Itoa(prNumber), "attachments", strconv.Itoa(attachmentID))
+	return s.svc.Download(ctx, spath)
+}
+
+// ──────────────────────────────────────────────────────────────
+//  Constructor
+// ──────────────────────────────────────────────────────────────
+
+// NewAttachmentService creates and returns a new pullrequest AttachmentService.
+func NewAttachmentService(method *core.Method) *AttachmentService {
+	return &AttachmentService{svc: attachment.NewService(method)}
+}
