+++ internal/wiki/service_attachment.go
@@ -1,0 +1,83 @@
+package wiki
+
+import (
+	"context"
+	"path"
+	"strconv"
+
+	"github.com/nattokin/go-backlog/internal/attachment"
+	"github.com/nattokin/go-backlog/internal/core"
+	"github.com/nattokin/go-backlog/internal/model"
+	"github.com/nattokin/go-backlog/internal/validate"
+)
+
+// AttachmentService handles attachment-related Backlog API calls for wiki pages.
+// It delegates all HTTP operations to the shared attachment.Service and is
+// responsible only for validation and spath construction.
+type AttachmentService struct {
+	svc *attachment.Service
+}
+
+// Attach attaches files uploaded to the space to the specified wiki.
+//
+// Backlog API docs: https://developer.nulab.com/docs/backlog/api/2/attach-file-to-wiki
+func (s *AttachmentService) Attach(ctx context.Context, wikiID int, attachmentIDs []int) ([]*model.Attachment, error) {
+	if err := validate.ValidateWikiID(wikiID); err != nil {
+		return nil, err
+	}
+
+	spath := path.Join("wikis", strconv.Itoa(wikiID), "attachments")
+	return s.svc.Attach(ctx, spath, attachmentIDs)
+}
+
+// List returns a list of files attached to the wiki page.
+//
+// Backlog API docs: https://developer.nulab.com/docs/backlog/api/2/get-list-of-wiki-attachments
+func (s *AttachmentService) List(ctx context.Context, wikiID int) ([]*model.Attachment, error) {
+	if err := validate.ValidateWikiID(wikiID); err != nil {
+		return nil, err
+	}
+
+	spath := path.Join("wikis", strconv.Itoa(wikiID), "attachments")
+	return s.svc.List(ctx, spath)
+}
+
+// Remove removes an attachment from the wiki page.
+//
+// Backlog API docs: https://developer.nulab.com/docs/backlog/api/2/remove-wiki-attachment
+func (s *AttachmentService) Remove(ctx context.Context, wikiID, attachmentID int) (*model.Attachment, error) {
+	if err := validate.ValidateWikiID(wikiID); err != nil {
+		return nil, err
+	}
+	if err := validate.ValidateAttachmentID(attachmentID); err != nil {
+		return nil, err
+	}
+
+	spath := path.Join("wikis", strconv.Itoa(wikiID), "attachments", strconv.Itoa(attachmentID))
+	return s.svc.Remove(ctx, spath)
+}
+
+// Download downloads a file attached to the wiki page.
+// The caller is responsible for closing FileData.Body after use.
+//
+// Backlog API docs: https://developer.nulab.com/docs/backlog/api/2/get-wiki-page-attachment
+func (s *AttachmentService) Download(ctx context.Context, wikiID, attachmentID int) (*model.FileData, error) {
+	if err := validate.ValidateWikiID(wikiID); err != nil {
+		return nil, err
+	}
+	if err := validate.ValidateAttachmentID(attachmentID); err != nil {
+		return nil, err
+	}
+
+	spath := path.Join("wikis", strconv.Itoa(wikiID), "attachments", strconv.Itoa(attachmentID))
+	return s.svc.Download(ctx, spath)
+}
+
+// ──────────────────────────────────────────────────────────────
+//  Constructor
+// ──────────────────────────────────────────────────────────────
+
+// NewAttachmentService creates and returns a new wiki AttachmentService.
+func NewAttachmentService(method *core.Method) *AttachmentService {
+	return &AttachmentService{svc: attachment.NewService(method)}
+}
